Correct notice
