+++ BeatMaker/Presentation/UsedSoundsList/UsedSoundsListView.swift
@@ -35,7 +35,7 @@ struct UsedTreckView<ViewModel: UsedTreckViewModeling>: View {
             ScrollView {
                 ForEach(viewModel.state.usedSoundsArray, id: \.self) { track in
                     HStack {
-                        trackView(sound: track.sound ?? Sound(id: "0", audioFileId: nil, name: "hype"))
+                        TrackView(sound: track.sound ?? Sound(id: "0", audioFileId: nil, name: "hype"))
                         
                         ZStack {
                             RoundedRectangle(cornerRadius: 5)
@@ -70,7 +70,6 @@ struct UsedTreckView<ViewModel: UsedTreckViewModeling>: View {
                         }
                         
                     }
-                    .cornerRadius(10)
                     .onTapGesture {
                         
                     }
